+++ NEWS.md
@@ -1,6 +1,17 @@
 Revision history for the R/broman package
 -----------------------------------------
 
+## Version 0.94, 2026-06-08
+
+- Added argument `notexact` to `crayons()`. If TRUE, return all
+  partial matches to any of the provided strings, ignoring cases.
+
+- `plot_crayons()` now takes a `colors` argument; give it a list of
+  crayon colors and it will show just that subset.
+
+- Add package logo.
+
+
 ## Version 0.92, 2025-12-08
 
 - Removed includes of `R_ext/PrtUtil.h` which began causing warnings on CRAN.